@@ -44,7 +44,7 @@ abstract class GeocodingPlatform extends PlatformInterface {
   /// [languageCode]_[countryCode] (eg. en_US or nl_NL).
   Future<List<Location>> locationFromAddress(
     String address, {
-    String localeIdentifier,
+    String? localeIdentifier,
   }) {
     throw UnimplementedError(
         'locationFromAddress() has not been implementated.');
@@ -65,7 +65,7 @@ abstract class GeocodingPlatform extends PlatformInterface {
   Future<List<Placemark>> placemarkFromCoordinates(
     double latitude,
     double longitude, {
-    String localeIdentifier,
+    String? localeIdentifier,
   }) {
     throw UnimplementedError(
         'placemarkFromCoordinates() has not been implementated.');
